Refactor MDX types out of Case Study files
