+++ packages/web/docs/src/app/case-studies/case-study-types.ts
@@ -1,4 +1,4 @@
-import type { getPageMap } from '@theguild/components/server';
+import type { MdxFile } from '../../mdx-types';
 
 export type CaseStudyFrontmatter = {
   title: string;
@@ -13,18 +13,4 @@ export type CaseStudyAuthor = {
   avatar?: string;
 };
 
-/**
- * TODO: This type should be exported from `nextra` and `@theguild/components`
- */
-export type MdxFile<FrontMatterType> = {
-  name: string;
-  route: string;
-  frontMatter?: FrontMatterType;
-};
-
 export type CaseStudyFile = Required<MdxFile<CaseStudyFrontmatter>>;
-
-/**
- * TODO: This should be exported from `nextra` and `@theguild/components`
- */
-export type PageMapItem = Awaited<ReturnType<typeof getPageMap>>[number];
